+++ packages/compliance-controller/CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `ComplianceService` support for an explicit Compliance API URL.
+- Add `selectAreAnyWalletsBlocked`.
+
 ### Changed
 
 - Bump `@metamask/controller-utils` from `^12.0.0` to `^12.1.0` ([#8774](https://github.com/MetaMask/core/pull/8774))
+
+### Fixed
+
+- Match EVM address casing consistently when reading cached wallet compliance statuses.
 
 ## [2.0.1]